refactor: move the "transcript" from "halo2_common" to "halo2_backend"
