+++ halo2_backend/src/lib.rs
@@ -2,9 +2,9 @@ pub mod arithmetic;
 mod helpers;
 pub mod plonk;
 pub mod poly;
+pub mod transcript;
 
 // Internal re-exports
 pub use halo2_common::circuit;
 pub use halo2_common::multicore;
-pub use halo2_common::transcript;
 pub use halo2_common::SerdeFormat;
